+++ samply-debugid/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "samply-debugid"
+version = "0.1.0"
+authors = ["Markus Stange <mstange.moz@gmail.com>"]
+license = "MIT OR Apache-2.0"
+edition = "2021"
+description = "Samply compatible debugids."
+repository = "https://github.com/mstange/samply/"
+readme = "README.md"
+
+[dependencies.debugid]
+default-features = false
+version = "0.8.0"
+
+[dependencies.object]
+default-features = false
+version = "0.37"
+
+[dependencies.uuid]
+default-features = false
+version = "1"
