+++ samply-debugid/README.md
@@ -1,0 +1,5 @@
+# samply-debugid
+
+This crate allows generating [`debugid`s](https://crates.io/crates/debugid)
+that are compatible with `samply`. Useful for writing your own profiles to be
+symbolicated and displayed with `samply load`.
