+++ samply-debugid/src/lib.rs
@@ -1,0 +1,5 @@
+mod codeid;
+mod debugid;
+
+pub use codeid::{CodeId, ElfBuildId, PeCodeId};
+pub use debugid::{code_id_for_object, debug_id_for_object, DebugIdExt};
