+++ samply-symbols/src/lib.rs
@@ -211,6 +211,7 @@ use jitdump::JitDumpIndex;
 use linux_perf_data::jitdump::JitDumpReader;
 use object::read::FileKind;
 pub use pdb_addr2line::pdb;
+pub use samply_debugid::{CodeId, ElfBuildId, PeCodeId};
 use shared::FileContentsCursor;
 pub use {debugid, object};
 
@@ -254,11 +255,11 @@ pub use crate::macho::FatArchiveMember;
 pub use crate::mapped_path::MappedPath;
 pub use crate::path_interner::PathInterner;
 pub use crate::shared::{
-    relative_address_base, AddressInfo, CandidatePathInfo, CodeId, ElfBuildId,
-    ExternalFileAddressInFileRef, ExternalFileAddressRef, ExternalFileRef, FileAndPathHelper,
-    FileAndPathHelperError, FileAndPathHelperResult, FileContents, FileContentsWrapper,
-    FileLocation, FrameDebugInfo, FramesLookupResult, LibraryInfo, LookupAddress,
-    MultiArchDisambiguator, OptionallySendFuture, PeCodeId, SymbolInfo, SyncAddressInfo,
+    relative_address_base, AddressInfo, CandidatePathInfo, ExternalFileAddressInFileRef,
+    ExternalFileAddressRef, ExternalFileRef, FileAndPathHelper, FileAndPathHelperError,
+    FileAndPathHelperResult, FileContents, FileContentsWrapper, FileLocation, FrameDebugInfo,
+    FramesLookupResult, LibraryInfo, LookupAddress, MultiArchDisambiguator, OptionallySendFuture,
+    SymbolInfo, SyncAddressInfo,
 };
 pub use crate::source_file_path::{SourceFilePath, SourceFilePathHandle, SourceFilePathIndex};
 pub use crate::symbol_map::{AccessPatternHint, SymbolMap, SymbolMapTrait};
